+++ CloudShell.ControlPlane.Client/ControlPlaneClientServiceCollectionExtensions.cs
@@ -130,6 +130,8 @@ public static class ControlPlaneClientServiceCollectionExtensions
             serviceProvider => serviceProvider.GetRequiredService<IControlPlane>());
         services.AddScoped<IResourceDeploymentManager>(
             serviceProvider => serviceProvider.GetRequiredService<IControlPlane>());
+        services.AddScoped<IResourceReplicaSlotStateManager>(
+            serviceProvider => serviceProvider.GetRequiredService<IControlPlane>());
         services.AddScoped<ILogManager>(
             serviceProvider => serviceProvider.GetRequiredService<IControlPlane>());
         services.AddScoped<ITraceManager>(